feat: fuzzy catalog search with rapidfuzz (GH-250) (#565)
* docs: add design spec for reducing local imports


* docs: fix spec per review — correct counts, add missing imports

Reviewer found: wrong remaining count (~20 should be ~28), missing
files (quick_report_tools, typing imports, desktop model imports),
and unclear attribution of manager.py webgui imports.


* docs: add implementation plan for reducing local imports

Reviewed and fixed: dead imports after extraction, mypy TYPE_CHECKING
fallback for non-Windows, _flatten in _discovery.py, co-located import
warnings, git add -A → git add -u.


* docs: add spec for credential-based test gating


* feat: add fuzzy search to all 4 catalog search modules (GH-250)

Add rapidfuzz dependency for fuzzy string matching. When exact token
matching fails, fall back to rapidfuzz WRatio with score_cutoff=50.
This catches partial stems, typos, and word reordering.

Example: searching "Klärfälle" now finds EMMACL ("Klärungsliste
anzeigen") which previously returned no results.

Applied to: transactions, tables, classes, function modules.


* fix: address review — consistent float scores, fuzzy tests for all modules

- Change score type from int to float in tables/classes/FMs (was int
  with truncation, now float matching transaction module)
- Add fuzzy search tests for tables, classes, and FM modules
- Fix match_type comment to include "fuzzy"


* fix: add pylint disables for too-many-locals/branches in search functions

The fuzzy matching block pushed branch/variable counts over pylint
limits. Add inline disables on the 3 affected search functions.


* fix: type_check — use float consistently for scores in all search modules

Change score field from int to float in Pydantic result models
(TableSearchResultItem, FMSearchResultItem, ClassSearchResultItem)
and initialize loop variables as 0.0. Also add comment that fuzzy
intentionally only matches entity descriptions, not fields/params.


---------

+++ src/sapwebguimcp/tools/fm_tools.py
@@ -65,7 +65,7 @@ class FMSearchResultItem(BaseModel):
     is_rfc_enabled: bool = Field(default=False, description="RFC-enabled flag")
     import_params: list[FMParameterResult] = Field(description="Import parameters")
     export_params: list[FMParameterResult] = Field(description="Export parameters")
-    score: int = Field(description="Relevance score (0-100)")
+    score: float = Field(description="Relevance score (0-100)")
     match_reason: str = Field(description="Why this FM matched")
